Remove unused routes and split payment update into its own controller
